[deckhouse] turn the metadata mapping into dto methods
- `dto.ModuleDefinition` converts itself to the MPV status metadata via `ConvertToStatusMetadata`, next to the sibling `Convert`.
- The legacy module.yaml requirements projection lives in the module-package-version controller, beside `legacyAccessibilityToCR`.
- Both metadata writers (the controller and versionsync) call the same method, so equal input keeps giving equal metadata.
